Initial Django project
Creates an empty django project called opensafely_interactive. It can be
run using the `just run` command.

This includes running postgres via a docker container. The settings for
connecting to this are hard-coded in settings.py; this should be changed
before we deploy to a production-like environment.

There's a basic unit/integration test to check that everything is
working in test_admin.py. This should be iterated upon as we develop
actual views to test.