Use psycopg2-binary
The maintainers recommend building psycopg2 from source, using the
'psycopg2' package to avoid conflicts with the system supplied libssl
and libgq libs. However, in practice this hasn't been an issue for us,
not least because we're deploying with Docker. It would also
potentially cause issues for dev using M1 Macs. Therefore, we're
choosing to use the binary, in line with our other applications.