Create local dev and test envs using docker
This allows the tests to be run in CI against a copy of the docker
image we'll be using in production. It allows allows local development
and testing using docker.

The database service has been moved into the main docker compose file.
In order for the other services to connect to it, the database
connection URL has been overriden for dev and test environments.