Add custom User model
This will give us more flexibility in future for adding custom
behaviours, such as sending out set password links and welcome emails on
User creation. It also allows us to use our own custom user id, instead
of relying on an autoincrementing integer.

We need to inherit from the AbstractBaseUser because we aren't using the
username field and this causes problems otherwise (since we end up with
a null entry for a non-nullable database field). However, using the
PermissionsMixin means we don't need to implement the logic for creating
superusers and other permissions.

We also need to override some of the behaviour of the UserManager,
because we're using an email field instead of a username.