Refactor hand made opencodelists fixtures using responses.
The existing fixtures had some issues:

1) autouse meant your tests miss cases when you expected the codelists
   API call but it didn't actually happen.
2) Patching all `requests.get()` call to return codelists was just not
   going to work with the other calls we had to make.
3) The autouse deletion of `request.session.request` was *super*
   confusing, and had no way to undo it if needed.

I have refactored these fixtures to use the excellent responses
library from the sentry folks, which allows flexible per-url
overriding of API responses. The pytest-responses plugin provides
a handy fixture, as well as disabling un-mocked network access in tests,
but with an error message that indicates what has happened, and a way to
allow network access on a per test basis (e.g. functional tests)

Additionally, I moved the fixture definition close to the tests that use
it, and made it explicit, so that you get an error if you did mock it
but did not call it in your code under test.