Consolidate job-server urls
We previously added JOB_SERVER_URL, but we still had a separate
JOB_SERVER_JOBS_URL separately, so misconfiguration was possible.

This change consolidates all job server urls to use JOB_SERVER_URL as
their base, in tests as well.

Partly because it's used in multiple places, it moves the loading and
validation of the related env vars into settings.py, and makes it
a `furl` object, for easier use, and to be able to kill all use of
urljoin, which pleases me.

It converts the slack links to use it, and removes the separately set
JOB_SERVER_JOBS_URL.

To help maintain this behaviour in the future, it refactors the
`services/jobserver.py` to use a common client to provide a single place
to manage urls and handle auth. Similarly in the tests, it adds a more
general `add_jobserver_response` fixture, similar to the
`add_codelists_response`. This ensures the tests are using the same
JOB_SERVER_URL as the application code.