feat: add the codelist report code into the repo
This creates self contained report code in `reports/codelist`. In aims
to consolidate all the codelist report specific information in once
place. The goal for this is primarily to local all the related code in
once location, but also to support multiple OSI reports in the future.

As part of this, it cleans up the file generation, by getting rid of the
the awkward variables.py set up, and moving to using cli args in the
project.yaml

It adds a full integration test, which uses opensafely to run the
report, and imports and fixes the pre-existing tests.

There are a few tweaks to study code and tests made in concert with
Louis.