Reduce spamminess of emails
Recent changes to the from address has caused emails to be marked as spam,
because it now differs from the reply-to.

This change removes the reply-to field and adds some text to the email to let
users know how to contact us, since hitting reply won't work. The word "here"
has also been removed from just before a hyperlink, as that can also increase
our spam rating.

A link to the feedback form has also been included for users receiving html
emails.