Reduce spamminess of emails
Recent changes to the from address has caused emails to be marked as spam.

This change improves the display of the from and reply-to addresses. It also
makes some content improvements; the word "here" has also been removed from
just before a hyperlink, as that can also increase our spam rating, and a link
to the feedback form has been included for users receiving html emails.