Cope with GC pauses in Hypothesis test
This test usually takes 1-2 ms to run. Very occasionally (every 10k
runs or so) it takes ~300 ms. By default Hypothesis fails tests that
take longer than 200 ms to run. This setting disables that behaviour.

We experimented with turning off GC, which makes the problem go away,
but that seems like asking for trouble and the occasional test that
slow isn't actually a problem for us.