break apart act and assert phases of success unit test
assert_difference mixes the act and assert phases of this test.  pytest
is unable to apply its magic to the internals of the decorator so
debugging a failure inside the context (eg submitting a form with a
missing field) requires working backwards from unrelated code.

We also should not be checking a relative change in this test, as the
comment mentions this way of testing would hide leaking tests.