replace logged-in testing helper
We were only using this in one place so rather than have the abstraction
this brings things into the test itself.

I've also reworked the test to not couple an view which we hope will
always require a login to test this flow, but instead lead on Django's
updating of is_authenticated.